refactor(codegen): extract type-layout registration into layout.rs
Relocate the named-type layout machinery (~1.2k lines) out of the llvm.rs
god-module: record/enum/machine named-struct predeclaration and body fill,
tagged-union layout construction, the host data-layout string, the heap-owning
composite-return classifiers, the layout-vec runtime declaration/type helpers,
and the layout-descriptor / owned-element / channel-element witness pointer
emitters, plus the CollectionLayoutWitness it returns.

The shared RecordLayoutMap / MachineCodegenLayout context types stay in
crate::llvm (held by FnCtx); layout imports them and llvm.rs calls back via
crate::layout::*; runtime_abi's layout-descriptor calls are re-qualified. Drops
the now-dead MachineLayout import from llvm.rs. Every emitted .ll is byte-
identical (ll-corpus verify clean, native and wasm32).
